+++ Vault/src/utopia/vault/nosql/access/UniqueModelAccess.scala
@@ -2,8 +2,7 @@ package utopia.vault.nosql.access
 
 import utopia.flow.datastructure.immutable.Value
 import utopia.vault.database.Connection
-import utopia.vault.model.immutable.Column
-import utopia.vault.sql.{Update, Where}
+import utopia.vault.sql.Condition
 
 /**
  * Common trait for access points which target an individual and unique model.
@@ -11,8 +10,18 @@ import utopia.vault.sql.{Update, Where}
  * @author Mikko Hilpinen
  * @since 31.3.2021, v1.6.1
  */
-trait UniqueModelAccess[+A] extends SingleModelAccess[A] with UniqueAccess[A]
+trait UniqueModelAccess[+A] extends SingleModelAccess[A] with DistinctModelAccess[A, Option[A], Value]
 {
+	// ABSTRACT -------------------------
+	
+	/**
+	 * @return Condition defined by this access point
+	 */
+	def condition: Condition
+	
+	
+	// COMPUTED -------------------------
+	
 	/**
 	 * @param connection Database connection (implicit)
 	 * @return The index of this model in database (may be empty)
@@ -23,40 +32,8 @@ trait UniqueModelAccess[+A] extends SingleModelAccess[A] with UniqueAccess[A]
 		case None => Value.empty
 	}
 	
-	/**
-	 * Reads the value of an individual column in this model
-	 * @param column Column to read
-	 * @param connection DB Connection (implicit)
-	 * @return value of that column (may be empty)
-	 */
-	def pullColumn(column: Column)(implicit connection: Connection) = readColumn(column)
 	
-	/**
-	 * Reads the value of an individual attribute / column in this model
-	 * @param attributeName Name of the attribute to read
-	 * @param connection DB Connection (implicit)
-	 * @return value of that attribute (may be empty)
-	 */
-	def pullAttribute(attributeName: String)(implicit connection: Connection) =
-		pullColumn(table(attributeName))
+	// IMPLEMENTED  ---------------------
 	
-	/**
-	 * Updates the value of an individual column in this model
-	 * @param column Column to update
-	 * @param value Value to assign to the column
-	 * @param connection DB Connection (implicit)
-	 * @return Whether any row was updated
-	 */
-	def putColumn(column: Column, value: Value)(implicit connection: Connection) =
-		connection(Update(target, column, value) + Where(condition)).updatedRows
-	
-	/**
-	 * Updates the value of an individual attribute / column in this model
-	 * @param attributeName Name of the attribute / property which is updated
-	 * @param value Value to assign to the attribute
-	 * @param connection DB Connection (implicit)
-	 * @return Whether any row was updated
-	 */
-	def putAttribute(attributeName: String, value: Value)(implicit connection: Connection) =
-		putColumn(table(attributeName), value)
+	override def globalCondition = Some(condition)
 }